+++ backend/internal/handler/subject_handler.go
@@ -134,7 +134,7 @@ func (h *SubjectHandler) Update(w http.ResponseWriter, r *http.Request) {
 		helpers.BadRequestFieldError(h.log, w, "id")
 		return
 	}
-	subjectID := uint8(subjectID64)
+	subjectID := uint16(subjectID64)
 	// Assemble DTO (all fields are optional)
 	dto := service.UpdateSubjectDTO{}
 	if nameFields := r.PostForm["name"]; len(nameFields) == 1 {
@@ -169,7 +169,7 @@ func (h *SubjectHandler) Delete(w http.ResponseWriter, r *http.Request) {
 		helpers.BadRequestFieldError(h.log, w, "id")
 		return
 	}
-	subjectID := uint8(subjectID64)
+	subjectID := uint16(subjectID64)
 	// Delete subject
 	if err := h.subjectService.DeleteSubject(r.Context(), subjectID); err != nil {
 		helpers.HandleServiceError(h.log, w, fmt.Errorf("failed to delete the subject: %w", err))
