+++ backend/internal/model/institution_administrator_position.go
@@ -5,7 +5,7 @@ import (
 )
 
 type InstitutionAdministratorPosition struct {
-	ID        uint8 `gorm:"primaryKey"`
+	ID        uint16 `gorm:"primaryKey"`
 	CreatedAt time.Time
 	UpdatedAt time.Time
 	Name      string `gorm:"type:varchar(100);unique;check:length(trim(name)) >= 4"`
