+++ backend/internal/model/room.go
@@ -7,7 +7,7 @@ import (
 
 // Room provides model of table with list of rooms (cabinets, dining room, etc.)
 type Room struct {
-	ID        uint8 `gorm:"primaryKey"`
+	ID        uint16 `gorm:"primaryKey"`
 	CreatedAt time.Time
 	UpdatedAt time.Time
 	Name      string `gorm:"type:varchar(20);unique;check:length(trim(name)) >= 1"`
