+++ backend/internal/model/staff.go
@@ -15,6 +15,6 @@ type Staff struct {
 	CreatedAt time.Time
 	UpdatedAt time.Time
 
-	PositionID uint8
+	PositionID uint16
 	Position   StaffPosition `gorm:"foreignKey:PositionID;references:ID"`
 }
